[policy] Add policy-enforcement-configuration.
As the textual policy configuration is currently the same for all
devices, add the configuration here for now. If need for device specific
changes arise, the configuration can be added to
etc/pulse/xpolicy.conf.d directory. However, current
policy-enforcement-module only supports extending the stream section, so
if changes are needed configuration file loading needs to be extended as
well.