[configs] Add missing parameter modes.
Due to kind-of-bug in module-meego-parameters if mode definition is
missing and mode (or route) changes to missing mode, parameters module
doesn't report the mode change forward. This then causes bugs further
down the path as certain actions aren't done when mode changes.
Simple fix is to just introduce all modes in the configuration.