[policy] Add earpieceforcall for consistency. JB#36527
For consistency's sake earpieceforcall is going to be used in upcoming
policy-settings-common, so that all routes used during voice calls use
forcall suffix. Adding the card and device sections to xpolicy.conf
already now, as they do no harm if they exist before
policy-settings-common changes, but would break if they don't exist when
the update is in place.

Signed-off-by: Juho Hämäläinen <juho.hamalainen@jolla.com>