[configs] Load stream-restore based on bluez version. JB#38234
With bluez5 we don't have sink or source to control volume with, and so
we don't need sink-volume mode as it does nothing.

Load combine-sink for simultaneous playback to phone outputs and
bluetooth devices.

Load rescue-streams so that sink-inputs or source-outputs left hanging
after bluetooth device disappears are not killed but are gracefully
moved to new sink or source.

Signed-off-by: Juho Hämäläinen <juho.hamalainen@jolla.com>