Restart adbd if it exits without being stopped. JB#38630
When adbd needs to switch identity based on root/unroot commands it
makes an exit and expects to be restarted. In android this would be
handled via actions bound to related property changes. But current
systemd unit configuration leaves the adbd service in failed state.

Add restart configuration for adbd.service.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>