[policy] Add bthfp device config. JB#38564
Even though not currently used with bluez4, add bthfp device
configurations to the xpolicy.conf already.

With bluez5 policy configuration the old bthsp device is split to bthsp
and bthfp, to manage the two profiles easier. Old implementation only
had hwid property with bthsp device which contained the info about which
bluetooth profile was in use. When the two profiles have own devices in
policy configuration managing the routing is much simpler.