Relax systemCheck.d a bit on droid-hal-*
[systemCheck.d] Relax droid-hal-*. Contributes to JB#41325

When sharing an adaptation we can not be too specific on certain
packages such as droid-hal-* but we should refer to the provides
packages.

In the end we should think if we could add Requires lines to the
droid-hal-version package and require mainly that in this ha.check file,
but that is for future. :)

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>