configs: Add ability to use sparse specific to an android version
This adds the ability for a droid-configs package to use an android
version specific sparse in addition to the device and common sparse.
It sits between the common and the device sparse and can override the
common sparse if needed.

If android_version_major is not defined, this does nothing.

[configs] Add ability to use sparse specific to an android version. Fixes JB#52029

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>