sparse: Add a service for memory reclaim.
Perform a memory reclaim right after boot for all services
and once per day for system services.
The rationale is that android does a similar thing when use_compaction
is enabled. The idea is that memory that is unlikely to be needed again
will be put to swap. Of course we cannot know which memory will be
reused again but we let the kernel decide what to put back to RAM when
needed.

CONFIG_PROCESS_RECLAIM should be enabled, but this should already be
the case in android based adaptations.

[sparse] Add a service for memory reclaim. JB#57321