sparse: Disable update_verifier for ports starting from Android 10
update_verifier tries to verify Android updates which doesn't make
sense in the case of it running as droid-system.

update_verifier uses dm-verity to verify the update's state and will
restart when the verification fails.

This always fail in our case and will show up in the symptom of a boot loop.

Read here for more:
https://android.googlesource.com/platform/bootable/recovery/+/refs/heads/master/update_verifier
https://source.android.com/devices/tech/ota/ab
https://source.android.com/security/verifiedboot

[sparse] Disable update_verifier for ports starting from Android 10. JB#57317

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>