configs: Default to single output with twin outputs.
Most Android devices only support single output properly, and when
old implementation of pulseaudio-modules-droid was split to
legacy-jb2q, the new implementation dropped the feature altogether.

Also since we cannot have separate volume for wired output and speaker
with these modes, we would need to attenuate the volume anyway so that
we don't play the audio too loud if user has headset on.

In conclusion, at least with Android adaptations we should only use
single output route when dealing with wired accessories.

[configs] Default to single output with twin outputs. Fixes JB#61159